requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
youtube-transcript-api==0.6.1
openai-whisper==20231117
yt-dlp==2023.10.13
langchain
langchain-community
langchain-google-genai
langchain-groq
faiss-cpu==1.7.4
sentence-transformers==2.2.2
torch==2.1.0
numpy<2.0